noor-platform/backend/pyproject.toml
[tool.poetry]
name = "noor-platform-backend"
version = "7.1.0"
description = "NOOR Platform Backend - National Human Capital Intelligence System"
authors = ["NOOR Platform Team <team@noor.ae>"]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.4.0"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pymongo = "^4.5.0"
motor = "^3.3.1"
neo4j = "^5.14.0"
redis = "^5.0.0"
confluent-kafka = "^2.3.0"
langchain = "^0.0.340"
langchain-openai = "^0.0.2"
langchain-anthropic = "^0.0.1"
openai = "^1.3.0"
anthropic = "^0.7.0"
pinecone-client = "^2.2.4"
sentence-transformers = "^2.2.2"
xgboost = "^2.0.0"
lightgbm = "^4.1.0"
scikit-learn = "^1.3.2"
pandas = "^2.1.3"
numpy = "^1.26.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
httpx = "^0.25.1"
aiohttp = "^3.9.0"
pyjwt = "^2.8.0"
cryptography = "^41.0.7"
diffprivlib = "^0.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
isort = "^5.12.0"
bandit = "^1.7.5"
ipython = "^8.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=. --cov-report=html --cov-report=term-missing"
